Warning 제거 처리, 불필요한 파일 제거
[Delete]
x64/Debug/Engine_SOURCE.lib

Project : Engine_SOURCE
[Modify]
M1. hjFadeScript.cpp
- 랜덤 변수 초기화 시 사용되는 인자의 타입을 UINT타입으로 static_cast 처리

M2. hjGridScript.cpp
- 얻어온 윈도우 크기를 기반으로 선언 및 초기화 하는 width, height의 계산값 결과를 float 타입으로 static_cast 처리

M3. hjMath.h
- 재정의 되는 XM_CONSTEXPR Define 문 삭제

+++ Engine_SOURCE/hjFadeScript.cpp
@@ -121,7 +121,7 @@ namespace hj
 		auto curTime = std::chrono::system_clock::now();
 		auto duration = curTime.time_since_epoch();
 		auto millis = std::chrono::duration_cast<std::chrono::milliseconds>(duration).count();
-		std::mt19937 randomSeed(millis);
+		std::mt19937 randomSeed(static_cast<UINT>(millis));
 
 		std::uniform_real_distribution<float> valueRange(0.f, 1.1f);
