+++ Engine_SOURCE/hjGridScript.cpp
@@ -88,8 +88,8 @@ namespace hj
 
 		RECT winRect;
 		GetClientRect(application.GetHwnd(), &winRect);
-		float width = winRect.right - winRect.left;
-		float height = winRect.bottom - winRect.top;
+		float width = static_cast<float>(winRect.right - winRect.left);
+		float height = static_cast<float>(winRect.bottom - winRect.top);
 		Vector2 resolution(width, height);
 
 		// Constant Buffer
